add put method for message
