+++ app/messages/routes.py
@@ -1,69 +1,19 @@
-from fastapi import APIRouter, WebSocket, Depends, HTTPException, WebSocketDisconnect
+from fastapi import APIRouter, Depends, HTTPException
 from ..deps import get_db_connection
-from ..websocket_manager import ConnectionManager
-from .crud import insert_messages_to_db, insert_text_messages_to_db, get_few_messages_from_db
-from ..db_instance import db
+from .crud import update_last_read_timestamp
+from .crud import get_few_messages_from_db
+from .schemas import MessageReadStatus
 from ..utils.logger import logger
-import json
 import asyncpg
 from datetime import datetime
 
 
 
 router = APIRouter()
-manager = ConnectionManager()
-
-
 
 @router.get("/testing")
 async def Testing():
     return "hello this is messaging route"
-
-
-@router.websocket("/workspace/{workspace_id}/{user_id}")
-async def websocket_message_endpoint(websocket: WebSocket, workspace_id: int, user_id: int):
-    await manager.connect(workspace_id, websocket)
-
-    try:
-        while True:
-            data = await websocket.receive_text()
-            payload = json.loads(data)
-            
-            message_data = {
-                "workspace_id": workspace_id,
-                "sender_id": user_id,
-                "message_type": payload["type"],
-                "reply_to": payload["reply_to"] 
-            }
-
-            async with db.get_connection() as conn:
-                async with conn.transaction():
-                    message_id = await insert_messages_to_db(message_data, conn)
-
-                    if payload["type"] == "text":
-                        await insert_text_messages_to_db(text_data={"message_id": message_id, "content":payload["content"]}, conn=conn)
-                    
-
-            message_obj = {
-                "message_id": message_id,
-                "workspace_id": workspace_id,
-                "sender_id": user_id,
-                "type": payload["type"],
-                "content": payload.get("content"),
-                "reply_to": payload.get("reply_to"),
-                "sent_at": datetime.utcnow().isoformat()
-            }
-
-            await manager.broadcast(workspace_id, message_obj)
-
-    except WebSocketDisconnect as e:
-        logger.info(f"User {user_id} disconnects from workspace {workspace_id}")
-        manager.disconnect(workspace_id, websocket)
-    except Exception as e:
-        logger.info(f"Unexpected Error: {e}")
-        manager.disconnect(workspace_id, websocket)
-    finally:
-        manager.disconnect(workspace_id, websocket)
 
 
 @router.get("/workspace/{workspace_id}/messages")
@@ -74,3 +24,10 @@ async def get_messages(workspace_id: int, timestamp:datetime, conn: asyncpg.Conn
         return response
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Failed to retrieve messages from DB -> {e}")
+
+@router.put("/workspace/{workspace_id}/read-status/{user_id}")
+async def update_read_status(workspace_id:int, user_id:int, payload:MessageReadStatus, conn: asyncpg.Connection = Depends(get_db_connection)):
+    try:
+        return await update_last_read_timestamp(workspace_id, user_id, dict(payload), conn)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Failed to update read_status -> {e}")
